fix!: Add the many2one in update form
Originally, the many2one were not added to the update form because of
my single usecase. However it absolutely makes sense to be able to
modify the many2one relationships.

To keep the original behiaviour, you can use the `except_columns`
kwarg and add each of the `_id` ending fields.

BREAKING CHANGE: Now you have the many2one by default in the update
form.

Refs: #3