feat: return `submitted` state
Because we did not have access to the submitted state, manual control
flow was complicated. For instance, creating an object from an
`st.experimental_dialog` was impossible.

This can still be improved, but for now it already makes things more
manageable.

in the future, st_update_select_form and st_delete_select_form as well
as st_crud_tabs should return a boolean with the submit status, but it
is not strictly necessary for now because they internally call st.rerun
anyway.